feat: add calculate not initiaed, ongoing, concluded and delayed with real data

+++ src/Core/Domain/ResponsePlans/Handlers/ResponsePlansCommandHandler.cs
@@ -20,42 +20,50 @@ namespace Actions.Core.Domain.ResponsePlans.Handlers
             if (command != null && command.ResponsePlanList != null)
             {
                 var responsePlansDB = _repository.GetByMetadataId(command.MetadataId);
-                var responsePlansIdDB = responsePlansDB.Select(plan => plan.Id);
+                bool process = true;
+                if (command.ResponsePlanList.Count == 0 && (responsePlansDB == null || responsePlansDB.Count() == 0))
+                    process = false;
 
-                var listToAdd = command.ResponsePlanList
-                    .Where(responsePlan => !responsePlansIdDB.Contains(responsePlan.Id))
-                    .Select(responsePlan => new ResponsePlan(
-                        responsePlan.Description, 
-                        responsePlan.Responsible.Id, 
-                        responsePlan.DueDate,
-                        responsePlan.Status,
-                        responsePlan.ActualStartDate,
-                        responsePlan.ActualEndDate,
-                        responsePlan.Cost,
-                        responsePlan.Comments,
-                        command.MetadataId
-                    ));
-
-                _repository.InsertRange(listToAdd);
-
-                var responsePlansIds = command.ResponsePlanList.Select(plan => plan.Id);
-                var listToUpdate = responsePlansDB
-                    .Where(responsePlan => responsePlansIds.Contains(responsePlan.Id))
-                    .Select(responsePlan => ToResponsePlan(responsePlan));
-
-                foreach (var item in listToUpdate)
+                if (process)
                 {
-                    var responsePlan = command.ResponsePlanList.SingleOrDefault(plan => plan.Id == item.Id);
+                    var responsePlansIdDB = responsePlansDB.Select(plan => plan.Id);
 
-                    if (item.HasModified(responsePlan.Description, responsePlan.Responsible.Id, responsePlan.DueDate, responsePlan.OriginalDueDate, responsePlan.Status, responsePlan.ActualStartDate, responsePlan.ActualEndDate, responsePlan.Cost, responsePlan.Comments))
-                        _repository.Update(item);
+                    var listToAdd = command.ResponsePlanList
+                        .Where(responsePlan => !responsePlansIdDB.Contains(responsePlan.Id))
+                        .Select(responsePlan => new ResponsePlan(
+                            responsePlan.Description, 
+                            responsePlan.Responsible.Id, 
+                            responsePlan.DueDate,
+                            responsePlan.Status,
+                            responsePlan.ActualStartDate,
+                            responsePlan.ActualEndDate,
+                            responsePlan.Cost,
+                            responsePlan.Comments,
+                            command.MetadataId
+                        ));
+
+                    _repository.InsertRange(listToAdd);
+
+                    var responsePlansIds = command.ResponsePlanList.Select(plan => plan.Id);
+                    var listToUpdate = responsePlansDB
+                        .Where(responsePlan => responsePlansIds.Contains(responsePlan.Id))
+                        .Select(responsePlan => ToResponsePlan(responsePlan));
+
+                    foreach (var item in listToUpdate)
+                    {
+                        var responsePlan = command.ResponsePlanList.SingleOrDefault(plan => plan.Id == item.Id);
+
+                        if (item.HasModified(responsePlan.Description, responsePlan.Responsible.Id, responsePlan.DueDate, responsePlan.OriginalDueDate, responsePlan.Status, responsePlan.ActualStartDate, responsePlan.ActualEndDate, responsePlan.Cost, responsePlan.Comments))
+                            _repository.Update(item);
+                    }
+
+                    var listToDelete = responsePlansDB
+                        .Where(responsePlan => !responsePlansIds.Contains(responsePlan.Id))
+                        .Select(responsePlan => ToResponsePlan(responsePlan));
+
+                    _repository.DeleteRange(listToDelete);
+
                 }
-
-                var listToDelete = responsePlansDB
-                    .Where(responsePlan => !responsePlansIds.Contains(responsePlan.Id))
-                    .Select(responsePlan => ToResponsePlan(responsePlan));
-
-                _repository.DeleteRange(listToDelete);
             }
         }
